+++ src/main/php/functions.php
@@ -20,6 +20,18 @@ function throws(\Exception $e)
 }
 
 /**
+ * wraps given callable into another callable so that the given callable is returned and not executed
+ *
+ * @param   callable  $callable
+ * @return  callable
+ * @since   0.6.0
+ */
+function wrap(callable $callable)
+{
+    return function() use ($callable) { return $callable; };
+}
+
+/**
  * creates a list of invocation results
  *
  * @api