+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * fixed bug with optional array parameters
   * `onConsecutiveCalls()` now falls back to the default return value in case a method gets invoked more often than results are defined
   * fixed #4: `onConsecutiveCalls()` should allow callables which are executed
+  * fixed #2: returning a callable requires another callable by providing `bovigo\callmap\wrap()`
 
 
 0.5.0 (2015-04-12)